feat: 2. Basic annotations - floor

+++ python_variable_annotations/2-floor.py
@@ -4,6 +4,6 @@ float n as argument and returns the floor of the float.
 '''
 
 
-def floor(n: int) -> float:
+def floor(n: float) -> int:
     '''return the floor of the float'''
     return n
